Move renderers to module
