+++ Cargo.toml
@@ -16,3 +16,4 @@ getrandom = "0.2"
 randomize = "3"
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
+ratatui = "0.29.0"
