+++ src/renderer/mod.rs
@@ -1,0 +1,6 @@
+pub mod debug;
+pub mod ratatui;
+pub mod gui;
+
+pub use gui::run_gui;
+pub use debug::run_debug;
