+++ src/renderer/ratatui.rs
@@ -1,0 +1,1 @@
+// TODO: for the ratatui renderer
